+++ container-engine-lib/lib/backend_interface/objects/service/service_config_test.go
@@ -2,11 +2,12 @@ package service
 
 import (
 	"encoding/json"
-	"github.com/kurtosis-tech/kurtosis/container-engine-lib/lib/backend_interface/objects/image_registry_spec"
 	"testing"
 	"time"
 
 	"github.com/kurtosis-tech/kurtosis/container-engine-lib/lib/backend_interface/objects/image_build_spec"
+	"github.com/kurtosis-tech/kurtosis/container-engine-lib/lib/backend_interface/objects/image_download_mode"
+	"github.com/kurtosis-tech/kurtosis/container-engine-lib/lib/backend_interface/objects/image_registry_spec"
 	"github.com/kurtosis-tech/kurtosis/container-engine-lib/lib/backend_interface/objects/nix_build_spec"
 	"github.com/kurtosis-tech/kurtosis/container-engine-lib/lib/backend_interface/objects/port_spec"
 	"github.com/kurtosis-tech/kurtosis/container-engine-lib/lib/backend_interface/objects/service_directory"
@@ -88,6 +89,7 @@ func getServiceConfigForTest(t *testing.T, imageName string) *ServiceConfig {
 		testServiceUser(),
 		testToleration(),
 		testNodeSelectors(),
+		testImageDownloadMode(),
 	)
 	require.NoError(t, err)
 	return serviceConfig
@@ -183,7 +185,8 @@ func testImageBuildSpec() *image_build_spec.ImageBuildSpec {
 	return image_build_spec.NewImageBuildSpec(
 		"test-image",
 		"path",
-		"")
+		"",
+		nil)
 }
 
 func testImageRegistrySpec() *image_registry_spec.ImageRegistrySpec {
@@ -215,4 +218,8 @@ func testNodeSelectors() map[string]string {
 	return map[string]string{
 		"disktype": "ssd",
 	}
+}
+
+func testImageDownloadMode() image_download_mode.ImageDownloadMode {
+	return image_download_mode.ImageDownloadMode_Missing
 }