+++ core/server/api_container/server/startosis_engine/kurtosis_instruction/add_service/add_services.go
@@ -3,6 +3,7 @@ package add_service
 import (
 	"context"
 	"fmt"
+	"github.com/kurtosis-tech/kurtosis/container-engine-lib/lib/backend_interface/objects/image_download_mode"
 	"github.com/kurtosis-tech/kurtosis/container-engine-lib/lib/backend_interface/objects/service"
 	"github.com/kurtosis-tech/kurtosis/core/server/api_container/server/service_network"
 	"github.com/kurtosis-tech/kurtosis/core/server/api_container/server/startosis_engine/enclave_plan_persistence"
@@ -12,6 +13,7 @@ import (
 	"github.com/kurtosis-tech/kurtosis/core/server/api_container/server/startosis_engine/kurtosis_starlark_framework/builtin_argument"
 	"github.com/kurtosis-tech/kurtosis/core/server/api_container/server/startosis_engine/kurtosis_starlark_framework/kurtosis_plan_instruction"
 	"github.com/kurtosis-tech/kurtosis/core/server/api_container/server/startosis_engine/kurtosis_types/service_config"
+	"github.com/kurtosis-tech/kurtosis/core/server/api_container/server/startosis_engine/plan_yaml"
 	"github.com/kurtosis-tech/kurtosis/core/server/api_container/server/startosis_engine/runtime_value_store"
 	"github.com/kurtosis-tech/kurtosis/core/server/api_container/server/startosis_engine/startosis_constants"
 	"github.com/kurtosis-tech/kurtosis/core/server/api_container/server/startosis_engine/startosis_errors"
@@ -39,7 +41,8 @@ func NewAddServices(
 	packageId string,
 	packageContentProvider startosis_packages.PackageContentProvider,
 	packageReplaceOptions map[string]string,
-	interpretationTimeValueStore *interpretation_time_value_store.InterpretationTimeValueStore) *kurtosis_plan_instruction.KurtosisPlanInstruction {
+	interpretationTimeValueStore *interpretation_time_value_store.InterpretationTimeValueStore,
+	imageDownloadMode image_download_mode.ImageDownloadMode) *kurtosis_plan_instruction.KurtosisPlanInstruction {
 	return &kurtosis_plan_instruction.KurtosisPlanInstruction{
 		KurtosisBaseBuiltin: &kurtosis_starlark_framework.KurtosisBaseBuiltin{
 			Name: AddServicesBuiltinName,
@@ -71,9 +74,10 @@ func NewAddServices(
 				serviceConfigs:               nil, // populated at interpretation time
 				interpretationTimeValueStore: interpretationTimeValueStore,
 
-				resultUuids:     map[service.ServiceName]string{}, // populated at interpretation time
-				readyConditions: nil,                              // populated at interpretation time
-				description:     "",                               // populated at interpretation time
+				resultUuids:       map[service.ServiceName]string{}, // populated at interpretation time
+				readyConditions:   nil,                              // populated at interpretation time
+				description:       "",                               // populated at interpretation time
+				imageDownloadMode: imageDownloadMode,
 			}
 		},
 
@@ -99,6 +103,8 @@ type AddServicesCapabilities struct {
 
 	resultUuids map[service.ServiceName]string
 	description string
+
+	imageDownloadMode image_download_mode.ImageDownloadMode
 }
 
 func (builtin *AddServicesCapabilities) Interpret(locatorOfModuleInWhichThisBuiltInIsBeingCalled string, arguments *builtin_argument.ArgumentValuesSet) (starlark.Value, *startosis_errors.InterpretationError) {
@@ -113,6 +119,7 @@ func (builtin *AddServicesCapabilities) Interpret(locatorOfModuleInWhichThisBuil
 		builtin.packageId,
 		builtin.packageContentProvider,
 		builtin.packageReplaceOptions,
+		builtin.imageDownloadMode,
 	)
 	if interpretationErr != nil {
 		return nil, interpretationErr
@@ -363,6 +370,12 @@ func (builtin *AddServicesCapabilities) allServicesReadinessCheck(
 	return failedServiceChecksRegularMap
 }
 
+func (builtin *AddServicesCapabilities) UpdatePlan(plan *plan_yaml.PlanYaml) error {
+	// TOOD: Implement
+	logrus.Warn("ADD SERVICES NOT IMPLEMENTED YET FOR UPDATING PLAN YAML.")
+	return nil
+}
+
 func (builtin *AddServicesCapabilities) Description() string {
 	return builtin.description
 }
@@ -416,6 +429,7 @@ func validateAndConvertConfigsAndReadyConditions(
 	packageId string,
 	packageContentProvider startosis_packages.PackageContentProvider,
 	packageReplaceOptions map[string]string,
+	imageDownloadMode image_download_mode.ImageDownloadMode,
 ) (
 	map[service.ServiceName]*service.ServiceConfig,
 	map[service.ServiceName]*service_config.ReadyCondition,
@@ -444,7 +458,7 @@ func validateAndConvertConfigsAndReadyConditions(
 		if !isDictValueAServiceConfig {
 			return nil, nil, startosis_errors.NewInterpretationError("One value of the '%s' dictionary is not a ServiceConfig (was '%s'). Values of this argument should correspond to the config of the service to be added", ConfigsArgName, reflect.TypeOf(dictValue))
 		}
-		apiServiceConfig, interpretationErr := serviceConfig.ToKurtosisType(serviceNetwork, locatorOfModuleInWhichThisBuiltInIsBeingCalled, packageId, packageContentProvider, packageReplaceOptions)
+		apiServiceConfig, interpretationErr := serviceConfig.ToKurtosisType(serviceNetwork, locatorOfModuleInWhichThisBuiltInIsBeingCalled, packageId, packageContentProvider, packageReplaceOptions, imageDownloadMode)
 		if interpretationErr != nil {
 			return nil, nil, interpretationErr
 		}