+++ core/server/api_container/server/startosis_engine/kurtosis_instruction/upload_files/upload_files.go
@@ -10,10 +10,11 @@ import (
 	"github.com/kurtosis-tech/kurtosis/core/server/api_container/server/startosis_engine/kurtosis_starlark_framework"
 	"github.com/kurtosis-tech/kurtosis/core/server/api_container/server/startosis_engine/kurtosis_starlark_framework/builtin_argument"
 	"github.com/kurtosis-tech/kurtosis/core/server/api_container/server/startosis_engine/kurtosis_starlark_framework/kurtosis_plan_instruction"
+	"github.com/kurtosis-tech/kurtosis/core/server/api_container/server/startosis_engine/plan_yaml"
 	"github.com/kurtosis-tech/kurtosis/core/server/api_container/server/startosis_engine/startosis_errors"
 	"github.com/kurtosis-tech/kurtosis/core/server/api_container/server/startosis_engine/startosis_packages"
 	"github.com/kurtosis-tech/kurtosis/core/server/api_container/server/startosis_engine/startosis_validator"
-	"github.com/kurtosis-tech/kurtosis/path_compression"
+	"github.com/kurtosis-tech/kurtosis/path-compression"
 	"github.com/kurtosis-tech/stacktrace"
 	"go.starlark.net/starlark"
 	"os"
@@ -216,6 +217,14 @@ func (builtin *UploadFilesCapabilities) FillPersistableAttributes(builder *encla
 	).AddFilesArtifact(
 		builtin.artifactName, builtin.filesArtifactMd5,
 	)
+}
+
+func (builtin *UploadFilesCapabilities) UpdatePlan(plan *plan_yaml.PlanYaml) error {
+	err := plan.AddUploadFiles(builtin.artifactName, builtin.src)
+	if err != nil {
+		return stacktrace.Propagate(err, "An error occurred updating plan with upload files.")
+	}
+	return nil
 }
 
 func (builtin *UploadFilesCapabilities) Description() string {