+++ core/server/api_container/server/startosis_engine/plan_yaml/plan_yaml.go
@@ -1,17 +1,14 @@
-package startosis_engine
+package plan_yaml
+
+import "sort"
 
 const (
-	HTTP ApplicationProtocol = "HTTP"
-	UDP  TransportProtocol   = "UDP"
-	TCP  TransportProtocol   = "TCP"
-
-	SHELL  TaskType = "sh"
-	PYTHON TaskType = "python"
-	EXEC   TaskType = "exec"
+	shell  TaskType = "sh"
+	python TaskType = "python"
+	exec   TaskType = "exec"
 )
 
-// TODO: there's really no point in making any of these references, consider just making them copies
-type PlanYaml struct {
+type privatePlanYaml struct {
 	PackageId      string           `yaml:"packageId,omitempty"`
 	Services       []*Service       `yaml:"services,omitempty"`
 	FilesArtifacts []*FilesArtifact `yaml:"filesArtifacts,omitempty"`
@@ -20,16 +17,21 @@ type PlanYaml struct {
 
 // Service represents a service in the system.
 type Service struct {
-	Uuid       string                 `yaml:"uuid,omitempty"`       // done
-	Name       string                 `yaml:"name,omitempty"`       // done
-	Image      *ImageSpec             `yaml:"image,omitempty"`      // done
-	Cmd        []string               `yaml:"command,omitempty"`    // done
-	Entrypoint []string               `yaml:"entrypoint,omitempty"` // done
-	EnvVars    []*EnvironmentVariable `yaml:"envVars,omitempty"`    // done
-	Ports      []*Port                `yaml:"ports,omitempty"`      // done
-	Files      []*FileMount           `yaml:"files,omitempty"`      // done
+	Uuid       string                 `yaml:"uuid,omitempty"`
+	Name       string                 `yaml:"name,omitempty"`
+	Image      *ImageSpec             `yaml:"image,omitempty"`
+	Cmd        []string               `yaml:"command,omitempty"`
+	Entrypoint []string               `yaml:"entrypoint,omitempty"`
+	EnvVars    []*EnvironmentVariable `yaml:"envVars,omitempty"`
+	Ports      []*Port                `yaml:"ports,omitempty"`
+	Files      []*FileMount           `yaml:"files,omitempty"`
+}
 
-	// TODO: support remaining fields in the ServiceConfig
+func (s *Service) MarshalYAML() (interface{}, error) {
+	sort.Slice(s.EnvVars, func(i, j int) bool {
+		return s.EnvVars[i].Key < s.EnvVars[j].Key
+	})
+	return s, nil
 }
 
 type ImageSpec struct {
@@ -48,6 +50,13 @@ type FilesArtifact struct {
 	Uuid  string   `yaml:"uuid,omitempty"`
 	Name  string   `yaml:"name,omitempty"`
 	Files []string `yaml:"files,omitempty"`
+}
+
+func (f *FilesArtifact) MarshalYAML() (interface{}, error) {
+	sort.Slice(f.Files, func(i, j int) bool {
+		return f.Files[i] < f.Files[j]
+	})
+	return f, nil
 }
 
 // EnvironmentVariable represents an environment variable.
@@ -84,13 +93,13 @@ type Task struct {
 	TaskType TaskType         `yaml:"taskType,omitempty"` // done
 	RunCmd   []string         `yaml:"command,omitempty"`  // done
 	Image    string           `yaml:"image,omitempty"`    // done
-	Files    []*FileMount     `yaml:"files,omitempty"`    // done
-	Store    []*FilesArtifact `yaml:"store,omitempty"`    // done
+	Files    []*FileMount     `yaml:"files,omitempty"`
+	Store    []*FilesArtifact `yaml:"store,omitempty"`
 
-	// only exists on SHELL tasks
+	// only exists on shell tasks
 	EnvVars []*EnvironmentVariable `yaml:"envVar,omitempty"` // done
 
-	// only exists on PYTHON tasks
+	// only exists on python tasks
 	PythonPackages []string `yaml:"pythonPackages,omitempty"`
 	PythonArgs     []string `yaml:"pythonArgs,omitempty"`
 
@@ -99,5 +108,5 @@ type Task struct {
 	AcceptableCodes []int64 `yaml:"acceptableCodes,omitempty"`
 }
 
-// TaskType represents the type of task (either PYTHON or SHELL)
+// TaskType represents the type of task (either python or shell)
 type TaskType string